requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
dataforseo-client==1.0.34
frozenlist==1.4.1
idna==3.10
importlib_metadata==8.5.0
lxml==5.3.0
macholib==1.16.3
multidict==6.1.0
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.0.7
yarl==1.13.1
zipp==3.20.2
